% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildVolcanoData.R
\name{BuildVolcanoData}
\alias{BuildVolcanoData}
\title{Create The Mutation Plot}
\usage{
BuildVolcanoData(
  volcanoClass,
  SigName = "padj",
  FCName = "log2foldchange",
  pThres = 0.05,
  fcThres = 2
)
}
\arguments{
\item{volcanoClass}{object of class VolcanoClass}

\item{SigName}{Name of the column with pvalues in (default: padj)}

\item{FCName}{Name of the logFC column (default: log2foldchange)}

\item{pThres}{p threshold (default: 0.05)}
}
\description{
Create The Mutation Plot
}
\examples{
BuildVolcanoData(volcanoClass, SigName="padj",FCName="log2foldchange", pThres=0.05, fcThres=2)
}
\keyword{Mutation}
\keyword{PLot}
\keyword{Volcano}
\keyword{Waterfall}
