% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\docType{data}
\name{ExampleDGElist}
\alias{ExampleDGElist}
\title{Create The Mutation Plot}
\format{An object of class \code{data.frame} with 15000 rows and 2 columns.}
\usage{
ExampleDGElist
}
\arguments{
\item{DGE}{data.frame of different gene expression}

\item{Names}{TRUE/FALSE not used currently (default: FALSE)}

\item{SigName}{Name of the column with pvalues in (default: padj)}

\item{FCName}{Name of the logFC column (default: log2foldchange)}

\item{pThres}{p threshold (default: 0.05)}

\item{PRINT}{TRUE/FALSE - print the final plot? (default: TRUE)}
}
\description{
Create The Mutation Plot
}
\examples{
volcanoPlot(DGE, Names=FALSE, SigName="padj",FCName="log2foldchange", pThres=0.05, fcThres=2, PRINT=TRUE)
}
\keyword{PLot}
\keyword{Volcano}
