% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Create The Mutation Plot}
\usage{
plotVolcano(
  volcanoClass,
  Names = FALSE,
  pThres = 0.05,
  fcThres = 2,
  PRINT = TRUE
)
}
\arguments{
\item{volcanoClass}{object of class VolcanoClass}

\item{Names}{TRUE/FALSE not used currently (default: FALSE)}

\item{pThres}{p threshold (default: 0.05)}

\item{PRINT}{TRUE/FALSE - print the final plot? (default: TRUE)}
}
\description{
Create The Mutation Plot
}
\examples{
mutationPlot(Data, "SampleName", "GeneName", "Info")
}
\keyword{PLot}
\keyword{Volcano}
