% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano-package.R
\docType{package}
\name{volcano-package}
\alias{volcano}
\alias{volcano-package}
\title{volcano: Georeference for Volcano}
\description{
Provides georeferencing of the 'volcano' data set, in New Zealand Map Grid coordinates.
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
