% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano.tif}
\alias{volcano.tif}
\title{Path to 'volcano.tif'}
\usage{
volcano.tif()
}
\value{
character string ('/path/to/volcano.tif')
}
\description{
Just a file path for testing with
}
\examples{
volcano.tif()
}
