% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoClass.R
\docType{class}
\name{VolcanoPlot-class}
\alias{VolcanoPlot-class}
\alias{VolcanoPlot}
\title{Description}
\arguments{
\item{DGEresults}{dataframe of differentially expressed genes}

\item{plots}{List of plots}

\item{sigGenes}{List of dataframes of different gene groups.}

\item{plotData}{Data used to generate the plots.}
}
\description{
Description
}
\examples{
Volcano <- VolcanoPlot()

}
\keyword{Plot}
\keyword{Volcano}
